Fix end-of-completion detection due to automatic move of the cursor
When triggering omni-completion after few letters have been entered, vim
is moving the cursor back to the beginning of the partially entered word.

(The details rely to the number returned by the first call to
omni-completion(), which calculate the column where the word start)

Vim is then moving back the cursor to that position when displaying the
omni-completion menu. This were then detected as the end-of-completion.
After this detection, vim receive ensime last messages which hide the
omni-completion menu.

This is easily fixed by setting the 'showCompletion' at appropriate
moment.