Avoid maintening the state of the quickfix list [#10]
Unfortunately, BufWinLeave is not triggered on the quickfix list, which
make it annoying to track if the window has been closed.

Therefore, simply remove all tracking of the state of the quickfix list,
and open each time is necessary.
This also remove the burden of tracking the tab number in which has been
opened.

Take the opportunity to fix properly the "non-styling" of the
quickfix list by forcing line number and cursor line to be off.